Forced from the team to remove non atomic support
