--- src/main/java/com/icthh/xm/ms/dashboard/service/dto/BulkOperationStatus.java
@@ -1,5 +1,0 @@
-package com.icthh.xm.ms.dashboard.service.dto;
-
-public enum BulkOperationStatus {
-    CREATED, UPDATED, DELETED, FAILED
-}
