+++ src/test/java/com/icthh/xm/ms/dashboard/web/rest/AtomicBulkDashboardResourceIntTest.java
@@ -1,16 +1,5 @@
 package com.icthh.xm.ms.dashboard.web.rest;
 
-import static com.icthh.xm.ms.dashboard.util.FileUtils.readAsString;
-import static java.util.Arrays.asList;
-import static org.mockito.MockitoAnnotations.initMocks;
-import static org.springframework.http.MediaType.APPLICATION_JSON;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.delete;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.put;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
-import static org.springframework.test.web.servlet.setup.MockMvcBuilders.standaloneSetup;
-
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.icthh.xm.commons.i18n.error.web.ExceptionTranslator;
 import com.icthh.xm.ms.dashboard.DashboardApp;
@@ -29,10 +18,23 @@ import org.springframework.security.test.context.support.WithMockUser;
 import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.test.web.servlet.MockMvc;
 
+import java.util.stream.Stream;
+
+import static com.icthh.xm.ms.dashboard.util.FileUtils.readAsString;
+import static java.util.Arrays.asList;
+import static junit.framework.TestCase.assertNull;
+import static org.mockito.MockitoAnnotations.initMocks;
+import static org.springframework.http.MediaType.APPLICATION_JSON;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.delete;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.put;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+import static org.springframework.test.web.servlet.setup.MockMvcBuilders.standaloneSetup;
+
 @RunWith(SpringRunner.class)
 @WithMockUser(authorities = "SUPER-ADMIN")
 @SpringBootTest(classes = {DashboardApp.class, SecurityBeanOverrideConfiguration.class})
-public class BulkDashboardResourceTest {
+public class AtomicBulkDashboardResourceIntTest {
 
     private final ObjectMapper mapper = new ObjectMapper();
 
@@ -78,61 +80,57 @@ public class BulkDashboardResourceTest {
 
     @Test
     @SneakyThrows
-    public void shouldCreateDashboards() {
+    public void shouldAtomicCreateDashboards() {
         httpMock.perform(post("/api/bulk/dashboards")
             .contentType(APPLICATION_JSON)
-            .content(readAsString("bulkCreateDashboards.json")))
-            .andExpect(content().json(readAsString("expectedBulkCreatedDashboards.json")))
+            .param("isAtomicProcessing", "true")
+            .content(readAsString("bulkAtomicCreateDashboards.json")))
             .andExpect(status().isOk());
     }
 
     @Test
     @SneakyThrows
-    public void shouldUpdateDashboards() {
+    public void shouldAtomicUpdateDashboards() {
         httpMock.perform(put("/api/bulk/dashboards")
             .contentType(APPLICATION_JSON)
-            .content(readAsString("bulkUpdateDashboards.json")))
-            .andExpect(content().json(readAsString("expectedBulkUpdatedDashboards.json")))
+            .param("isAtomicProcessing", "true")
+            .content(readAsString("bulkAtomicUpdateDashboards.json")))
             .andExpect(status().isOk());
     }
 
     @Test
     @SneakyThrows
-    public void shouldDeleteDashboards() {
+    public void shouldAtomicDeleteDashboards() {
         httpMock.perform(delete("/api/bulk/dashboards")
             .contentType(APPLICATION_JSON)
-            .content(readAsString("bulkDeleteDashboards.json")))
-            .andExpect(content().json(readAsString("expectedBulkDeletedDashboards.json")))
+            .content(readAsString("bulkAtomicDeleteDashboards.json")))
             .andExpect(status().isOk());
+
+        assertNulls(953L, 954L);
     }
 
     @Test
     @SneakyThrows
-    public void shouldFailCreateDashboards() {
+    public void shouldFailAtomicCreateDashboards() {
         httpMock.perform(post("/api/bulk/dashboards")
             .contentType(APPLICATION_JSON)
-            .content(readAsString("failCreateBulkDashboards.json")))
-            .andExpect(content().json(readAsString("expectedFailedCreateBulkDashboards.json")))
-            .andExpect(status().isOk());
+            .param("isAtomicProcessing", "true")
+            .content(readAsString("failBulkAtomicCreateDashboards.json")))
+            .andExpect(status().isInternalServerError());
     }
 
     @Test
     @SneakyThrows
-    public void shouldFailUpdateDashboards() {
+    public void shouldFailAtomicUpdateDashboards() {
         httpMock.perform(put("/api/bulk/dashboards")
             .contentType(APPLICATION_JSON)
-            .content(readAsString("failUpdateBulkDashboards.json")))
-            .andExpect(content().json(readAsString("expectedFailedUpdateBulkDashboards.json")))
-            .andExpect(status().isOk());
+            .param("isAtomicProcessing", "true")
+            .content(readAsString("failBulkAtomicUpdateDashboards.json")))
+            .andExpect(status().isInternalServerError());
     }
 
-    @Test
-    @SneakyThrows
-    public void shouldFailDeleteDashboards() {
-        httpMock.perform(delete("/api/bulk/dashboards")
-            .contentType(APPLICATION_JSON)
-            .content(readAsString("failDeleteBulkDashboards.json")))
-                .andExpect(content().json(readAsString("expectedFailedDeleteBulkDashboards.json")))
-            .andExpect(status().isOk());
+    void assertNulls(Long ...ids) {
+        Stream.of(ids)
+            .forEach(id-> assertNull(dashboardRepository.findOneById(id)));
     }
 }
