--- src/test/resources/expectedBulkDeletedDashboards.json
@@ -1,90 +1,0 @@
-[
-    {
-        "status": "DELETED",
-        "dashboardItem": {
-            "id": 953,
-            "name": "Name of dashboard",
-            "owner": "me",
-            "typeKey": "TYPE_KEY",
-            "isPublic": false,
-            "layout": {},
-            "config": {
-                "hidden": true,
-                "name": {
-                    "en": "Config 1",
-                    "ru": "Config 1",
-                    "uk": "Config 1"
-                },
-                "icon": "no-icon",
-                "orderIndex": 5,
-                "menu": {
-                    "group": {
-                        "key": "MENU_KEY",
-                        "name": {
-                            "en": "KEY",
-                            "uk": "KEY",
-                            "ru": "KEY"
-                        },
-                        "orderIndex": 1,
-                        "icon": "swap_horizontal_circle"
-                    }
-                },
-                "slug": "slug"
-            },
-            "widgets": [
-                {
-                    "id": null,
-                    "selector": "ext-tenant-mnp/request-details-widget",
-                    "name": "Widget",
-                    "isPublic": false,
-                    "config": {},
-                    "dashboard": null
-                }
-            ]
-        }
-    },
-    {
-        "status": "DELETED",
-        "dashboardItem": {
-            "id": 954,
-            "name": "Name of dashboard",
-            "owner": "me",
-            "typeKey": "TYPE_KEY",
-            "isPublic": false,
-            "layout": {},
-            "config": {
-                "hidden": true,
-                "name": {
-                    "en": "Config 1",
-                    "ru": "Config 1",
-                    "uk": "Config 1"
-                },
-                "icon": "no-icon",
-                "orderIndex": 5,
-                "menu": {
-                    "group": {
-                        "key": "MENU_KEY",
-                        "name": {
-                            "en": "KEY",
-                            "uk": "KEY",
-                            "ru": "KEY"
-                        },
-                        "orderIndex": 1,
-                        "icon": "swap_horizontal_circle"
-                    }
-                },
-                "slug": "slug"
-            },
-            "widgets": [
-                {
-                    "id": null,
-                    "selector": "ext-tenant-mnp/request-details-widget",
-                    "name": "Widget",
-                    "isPublic": false,
-                    "config": {},
-                    "dashboard": null
-                }
-            ]
-        }
-    }
-]
