--- src/test/resources/expectedFailedUpdateBulkDashboards.json
@@ -1,81 +1,0 @@
-[
-    {
-        "status": "UPDATED",
-        "dashboardItem": {
-            "id": 955,
-            "name": "Name of dashboard",
-            "owner": "me",
-            "typeKey": "TYPE_KEY",
-            "isPublic": false,
-            "config": {
-                "hidden": true,
-                "name": {
-                    "en": "Config 1",
-                    "ru": "Config 1",
-                    "uk": "Config 1"
-                },
-                "icon": "no-icon",
-                "orderIndex": 5,
-                "menu": {
-                    "group": {
-                        "key": "MENU_KEY",
-                        "name": {
-                            "en": "KEY",
-                            "uk": "KEY",
-                            "ru": "KEY"
-                        },
-                        "orderIndex": 1,
-                        "icon": "swap_horizontal_circle"
-                    }
-                },
-                "slug": "slug"
-            },
-            "widgets": [
-                {
-                    "selector": "ext-tenant-mnp/request-details-widget",
-                    "name": "Widget",
-                    "isPublic": false
-                }
-            ]
-        }
-    },
-    {
-        "status": "FAILED",
-        "dashboardItem": {
-            "name": "Name of dashboard",
-            "owner": "me",
-            "typeKey": "TYPE_KEY",
-            "isPublic": false,
-            "config": {
-                "hidden": true,
-                "name": {
-                    "en": "Config 1",
-                    "ru": "Config 1",
-                    "uk": "Config 1"
-                },
-                "icon": "no-icon",
-                "orderIndex": 5,
-                "menu": {
-                    "group": {
-                        "key": "MENU_KEY",
-                        "name": {
-                            "en": "KEY",
-                            "ru": "KEY",
-                            "uk": "KEY"
-                        },
-                        "orderIndex": 1,
-                        "icon": "swap_horizontal_circle"
-                    }
-                },
-                "slug": "slug"
-            },
-            "widgets": [
-                {
-                    "selector": "ext-tenant-mnp/request-details-widget",
-                    "name": "Widget",
-                    "isPublic": false
-                }
-            ]
-        }
-    }
-]
